backend/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
yt-dlp
python-dotenv
